snapshot doesn't build

+++ Modules/SwiftRTCuda/mathOps.cu
@@ -442,7 +442,7 @@ cudaError_t srtPowN(
     cudaStream_t stream
 ) {
     Cast2TensorDescriptorsA(paDesc, poDesc)
-    return select<PowN, float>(a, aDesc, n, out, oDesc, stream);
+    return select<PowN, int>(a, aDesc, n, out, oDesc, stream);
 }
 
 //------------------------------------------------------------------------------
@@ -455,7 +455,7 @@ cudaError_t srtRoot(
     cudaStream_t stream
 ) {
     Cast2TensorDescriptorsA(paDesc, poDesc)
-    return select<Root, float>(a, aDesc, n, out, oDesc, stream);
+    return select<Root, int>(a, aDesc, n, out, oDesc, stream);
 }
 
 //------------------------------------------------------------------------------
